backend/requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
openai>=1.12.0
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.118.0
pydantic>=2.6.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.9 